Fix Koreans crashing the bot

+++ PPTMonitor/MainForm.cs
@@ -184,27 +184,27 @@ namespace PPTMonitor {
         private int menuStartFrames = 0;
 
         private void runLogic() {
-            if (GameHelper.boardAddress(PPT, playerID) != 0x0 && GameHelper.OutsideMenu(PPT) && GameHelper.getBigFrameCount(PPT) != 0x0) {
-                if (numplayers < 2) {
-                    ScanTimer.Enabled = false;
+            if (GameHelper.OutsideMenu(PPT) && GameHelper.CurrentMode(PPT) == 4 && numplayers < 2) {
+                ScanTimer.Enabled = false;
 
-                    foreach (var process in Process.GetProcessesByName("puyopuyotetris")) {
-                        process.Kill();
-                    }
-
-                    Thread.Sleep(10000);
-
-                    Process.Start("steam://rungameid/546050");
-
-                    Thread.Sleep(15000);
-
-                    buttonRehook_Click(this, EventArgs.Empty);
-
-                    ScanTimer.Enabled = true;
-
-                    return;
+                foreach (var process in Process.GetProcessesByName("puyopuyotetris")) {
+                    process.Kill();
                 }
 
+                Thread.Sleep(10000);
+
+                Process.Start("steam://rungameid/546050");
+
+                Thread.Sleep(15000);
+
+                buttonRehook_Click(this, EventArgs.Empty);
+
+                ScanTimer.Enabled = true;
+
+                return;
+            }
+
+            if (GameHelper.boardAddress(PPT, playerID) != 0x0 && GameHelper.OutsideMenu(PPT) && GameHelper.getBigFrameCount(PPT) != 0x0) {
                 int piecesAddress = GameHelper.piecesAddress(PPT, playerID);
 
                 int[] pieces = new int[5];
